python: move pysixtrack convert to a separate submodule of pysixtracklib
- Move BeamElement to pysixtrack_helper and rename it to BeamElementConverter
- Update pyisxtracklib_test.testdata to use the new class name and import
  path
- cosmetic changes (autopep8, etc.)
